requirements.txt
streamlit
pandas
numpy
matplotlib
ta
python-telegram-bot
MetaTrader5
schedule
python-dotenv
